requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiomisc==17.5.6
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
black==24.4.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.8.2
cron-descriptor==1.4.3
Django==5.0.4
django-celery-beat==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
frozenlist==1.4.1
idna==3.7
kombu==5.3.7
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
